// client/src/app/modules/collect-point/models/collect-point.ts
export interface CollectPoint {
    id: string
    cnpj: string
    //razaoSocial
    companyName: string
    //nomeFantasia
    tradingName: string
    //inscricaoEstadual
    stateRegister: number
    //cep
    zipCode: number
    street: string
    number: number
    neighborhood: string
    city: string
    state: string
    addressComplement: string
    registrationDate: string
    updateDate: string
}